requirements.txt
# It contains all packages that are required or useful when developing
# NaverWebtoonFeeds. If you just want to use the app, you don't have to install
# these packages. Type `python setup.py install` and you are ready.

Flask==0.10.1
Flask-Assets==0.9
Flask-Cache==0.12
Flask-gzip==0.1
Flask-Script==0.6.7
Flask-SQLAlchemy==1.0
requests==2.2.1
lxml==3.3.4
BeautifulSoup==3.2.1
PyYAML==3.11
yuicompressor==2.4.8
pytz==2014.2
netaddr==0.7.11
blinker==1.3

## Optional packages
MySQL-python==1.2.5
redis==2.9.1
rq==0.3.13
gunicorn==18.0

## Development & Testing
pytest==2.5.2
pytest-cov==1.6
pytest-pep8==1.0.5
mock==1.0.1
cssselect==0.9.1
tox==1.7.1
pylint==1.1.0
Sphinx==1.2.2
sphinx_rtd_theme==0.1.6
ipython==2.0.0
